use the right field name for diff-start
When we propose the final releases for milestone projects we set
diff-start to ensure the announcement email includes all of the
release notes for the deliverable. We were using the wrong field name,
which means the pike announcements did not include all of the
changes. This corrects the problem so we get the right content for
queens.

Change-Id: I7bf213d3f954381b0a556be0d310e5caad6cd18e
Signed-off-by: Doug Hellmann <doug@doughellmann.com>