Release os-testr 1.0.0
This commit releases os-testr 1.0.0, this is a major version bump
because it changes the testr runner used internally from testr to
stestr. This cleans up the code significantly and also means we stop
calling python via subprocess. However because stestr and testr aren't
100% compatible this has the potential to not be backwards compatible.
Since we're quite early in the queens cycle this is the best time to
make a release like this.

Change-Id: If9be85acf2454861e08e62ad4881e164bbec753f