new-branch: Add check for branch name before adding
If quickly scanning the command line arguments for the new-release
tool, it may not be obvious that --stable-branch is only needed
one time when initially creating the stable branch, not every time
doing a release off of a stable branch.

This adds a check to make sure the stable branch does not already
exist in the deliverable file before attempting to add the branch
directive for the current release.

Change-Id: Ib15aae2530eb9aebaad23d249dc3d59b022a6d46