switch from Popen to check_output in release notes tools
The copy of release_notes.py imported from the release-tools repo was
modified to use Popen instead of oslo.concurrency. However, that
change was made incorrectly, and an invalid exception was being
caught. When the commands work fine, this doesn't present a
problem. When the commands fail, it breaks error handling. Replace all
of the local logic with check_call(), the function used elsewhere in
this repo for running background commands. Update users of run_cmd()
in release_notes.py to only expect to receive stdout as a response.

Change-Id: I291a077ed613c3e4ad56b85c81945c15a6301c17
Signed-off-by: Doug Hellmann <doug@doughellmann.com>