make release job validation based on project type
Require puppet repositories to have the puppet jobs and node
repositories to have the node jobs.

Update the allowed values for release-type to include "puppet" and
"nodejs" for folks who want to be explicit, but also use the module
type-detection code to default properly.

Report any jobs that are valid release jobs but for the wrong release
type as errors instead of warnings.

Change-Id: I9330cc62834f42ae0cd4d1cc48ed963846d72944
Signed-off-by: Doug Hellmann <doug@doughellmann.com>