Release os-brick 2.0.0
This release includes a number of changes
to os-brick.  There are bugfixes as well as
improvements to multi-path support and a change
to avoid possible race conditions.

The release includes removal of legacy connector
constants that had been kept for backwards
compatibility.  As a result I am considering this
a backward incompatible change and bumping the
major version.

Change-Id: Ia852f69f1a2b9cdee02cad7ab9523ab959d71115