Release new oslo.rootwrap for Queens
The release contains a fix for a bug that should help neutron fullstack
gate. (The bug is when a timeout during a command execution leaves stale
data in the rootwrap daemon socket, and the data is then returned to the
next caller instead of being flushed.)

The minor version number is bumped because the release contains
requirements changes.

Change-Id: I87642f3105803ec2cc7bd3d5ac119c29db70697f
Related-Bug: #1654287