Proposed release schedule for Rocky
Assuming we will stay with a 6 month release cycle for Rocky,
this is a proposed schedule for the Rocky cycle.

One thing to note is the date for the S PTG has not been finalized
yet, so we may need or want to push out the window between one of
the milestones so there is not too long of a gap between Rocky
being released and the next PTG being held.

Change-Id: I0e26b443b7d4cf64f834b9726342cdfab54d36e7