Add doc job requirements.txt
We had previously installed our docs dependencies as extras in
setup.cfg. Doc jobs are now changed to not install the project,
so these dependencies are no longer being installed.

With recent changes to the doc jobs, we need to specificy that python3
is used. Update the zuul config file for this.

Change-Id: Ie925efefb1126dbe7fa0a9f0f1727da01f48c512
Depends-on: Ib882a4d09898ff206b36ac8aef840bee0748368e