tripleo-upgrade: create first tag / branch (independent)
This is a request from the TripleO Upgrades team.
They are working on this new project that run Ansible to control TripleO
upgrades.

They need a stable/pike branch, so this commit is an exceptional request
to create a branch on the latest commit, so the team can go ahead and
push some Queens only commits in master once the branch is created.

It also creates a 1.0.0 for Pike and later we'll create 2.0.0 for
Queens.

Note, this is an independent deliverable this time, because we missed
the Pike deadline.

Change-Id: I111b745fe44a899c24f8e088e4dbf46358ecfa56