os-vif: queens release 1.8.0
This is the first and so far only release for os-vif
in the queens series. Unless something major is needed
before the queens non-client library freeze, this will
also be the version we GA and branch for stable/queens.

This change also updates the release notes URL for the
pike deliverable.

Change-Id: Idd7f635295aff6e46391b82e45b89ef79d6d6c05