os-traits: queens release 0.5.0
This is most likely the GA version of os-traits for the queens
release and what we'd branch for stable/queens.

There are only 3 changes in this release. The minor version is
bumped because there are new minimum required package dependencies
in this release.

The release notes URL for the pike deliverable is also updated.

Change-Id: Ib0154ade88150523070bf8f63ace4fd23983b672