novaclient: queens release 10.0.0
This is the first but probably not the last release
for python-novaclient in the queens series. We still
have a few changes up for review for new microversion
support that we'd like to get in before queens feature
freeze, but we should get a release out now to make sure
it plays well with other services in the ecosystem.

Depends-On: I60490f3d74212bb172dccc1c7d337198e1021236

Change-Id: Ie402bae6bab0012d4faed342eaa7c6a0011b86e9