Switch cycle higlights to string list
The original thought was to have this contain rich formatting
like used in reno. But the reality is, for what this is used
for, we really just want a bulleted list of highlights for
each project.

Based on the first couple of patches, this has been a point of
confusion for those trying to use this directive. This patch
simplifies things by just treating the cycle-highlights as a
list that gets added as bullet items to the generated results.

Change-Id: If265606eb20fa3d49e3cf4a684ea683de4267927