Release 0.11.3 of openstacksdk for queens
The osc-functional-devstack-tips job, which tests master changes of
openstackclient and openstacksdk against each other with
openstackclient's functional tests was broken and was not testing
master against master but rather master of openstackclient against
released version of SDK. Therefore, the gate that was protecting us
against these breakages was incorrect.

The gate job has been fixed and SDK as of this commit fixes the
osc-functional-devstack-tips job. The osc-functional-devstack job, which
checks OSC against released SDK is broken with 0.11.2 because of the
bug fixed in this patch.

Change https://review.openstack.org/#/c/540554/ in OSC shows the -tips
jobs working. It depends on the master version of this fix. It also
shows the non-tips job broken, which is due to the issue here. When we
cut this release and bump the constraint, OSC's devstack job should go
green.

Change-Id: I1ff62751a3b62d050e1459c04931cb4d98681a48