deprecate the use of highlights attached to a release
We want teams to use reno for release notes and cycle-highlights for
more general notes. This patch disallows the use of release-specific
highlights for new releases from the new series with a message
explaining why. We can change the schema and let the yamllint perform
the check after running this way for a while.

Change-Id: I8e885b856f3412a3f6df691abbfd8c78037bcd91
Signed-off-by: Doug Hellmann <doug@doughellmann.com>