Release and cut queens branch for neutron-vpnaas and dashboard
In Queens release neutron-vpnaas and its dashboard use
the cycle-with-intermediary release model, but it is nice to cut
stable/queens at the same timing as for other neutron projects.

In addition, the i18n team will sets up stable/queens branch
on the translation platform around RC1 release, so it is helpful
if neutron-vpnaas-dashboard has a stable branch now.

Change-Id: Ie1e0592a22e2a97f60e49adf1583587925ed0e6a