Retry pypi name checking
We added logic to change underscores to dashes when checking package
names on pypi. Apparently there is a mix of conventions used here as
this caused some checks to fail.

This changes the check to try the expected sdist name first, and if
that fails, try again with the modified name.

Change-Id: I8c0f2cd6866f295a622c3ccce9620bf4358327af