Remove per-release highlights for Ocata
We've been warning that these are deprecated for awhile now, and
we now have enforcement in place to make sure no new highlights
are added.

We also have some automation that rewrites the deliverable files
but does not always do a great job, resulting in poorly formatted
text that causes warnings in the validation job.

Since many folks new to the release process, and some that have
done it for awhile, look at existing files to see what to do for
new releases, they do assume highlights should be used and end up
adding them to their deliverable file. Since most of these also
do not know to run local validatation, this results in gate
resources being used to just now be rejected with the current
checks.

To help prevent this from happening, and to get rid of the warning
messages from poor reformatting, this just removes highlights
from the last few releases.

Change-Id: I5753436c7fb99240eb7a6f691c81e580ef8e7ac9