Switch openstack-tox-validate to non-voting temporarily
The Puppet OpenStack project had a module which was not released
since 2014 and therefore it did not have an entry in the goverance
repository.  The current TC chair is not available so in order to
not delay the release, set the validate job to non-voting to skip
the governance check for release.

Change-Id: Idfea40c4f3a1d624f097df34cb53d59fe25b4aaa
Related-Change: I54c4a9e9b2ec988914bc5174a599ab17ffe60cdc