remove the 0.4.0.0b2 tag info for blazar
The releases site tries to show the first "full" release for a project
for a series, but because blazar changed numbering in the middle of the
cycle the logic doesn't know what to do with the old tag info and ends
up making it look like blazar started queens with a beta version. That
is different from other cycle-with-milestone projects, and is confusing,
so just discard the old information to fix the releases site.

Change-Id: I2861add115ff4b2a6740fa66f3100e247b332c23
Signed-off-by: Doug Hellmann <doug@doughellmann.com>