Release os-brick 2.4.0
This releases os-brick 2.4.0 to get a number of
bug fixes released and to also add a couple of new features.
The first feature is the ability for ScaleIO to extend
attached volumes.  It also adds support for SMBFS mounting
vhd/x images.

Given the feature changes this bumps os-brick from
2.3.0 to 2.4.0.

Change-Id: I2889476829629ac0494fc9ed339ddb4a5957438b