Tag nova-powervm 6.0.1
Requests a tag for nova-powervm 6.0.1.

This is the first tag request for nova-powervm in openstack/releases
so the yaml file did not previously exist. This adds some of the
previous branches and tags to that file as well for the sake of
tooling.

This is done in the _independent directory because nova-powervm
was not yet under governance in the Queens timeframe.

Change-Id: I90aa1a08301885fd262b501bb4f8e54632b97445