Tag networking-powervm 6.0.1
Requests a tag for networking-powervm 6.0.1.

This is the first tag request for networking-powervm in
openstack/releases so the yaml file did not previously exist. This
adds some of the previous branches and tags to that file as well
for the sake of tooling.

This is done in the _independent directory because
networking-powervm was not yet under governance in the Queens
timeframe.

Change-Id: Ic553bf3e9c138d026ccd4e2980dda064a729333c