ignore the reno cache when building the set of release notes
For list-changes and the announcement emails we want to ignore any
local cache file created when a package was built and scan the git
repository again. The cache file is meant to be used by the consumer
of the package being released, but we have the git repository so we
don't need it and ignoring it will ensure we have the correct notes in
any announcements.

Change-Id: I047f7e455a1d238d12cf94a60cd0d44f44406a3d
Signed-off-by: Doug Hellmann <doug@doughellmann.com>