Release python-openstackclient 3.14.2 (stable/queens)
This contains a change needed to enable novaclient (that supports the
volume multi-attach feature in queens) to work together with
openstackclient for the queens release.

Once openstackclient is released, we can propose an update of the
novaclient constraints from 9.1.1 => 10.1.0 to enable users to be able
to use the volume multi-attach feature in nova out-of-the-box when they
install the queens release of openstack.

Change-Id: Ic84cf5f4c7ea358a201c0e23cd4135dafbe680d8