update new-release to process all repositories
new-release was using the repositories tagged as part of the most
recent release to determine what to tag for the new release. This
change updates it to look at all of the repositories known to be part
of the deliverable to see which ones may need to be tagged. It tries
to report the last time a repository was tagged and still tries to
skip over repositories that do not need to be tagged because there are
no changes.

Change-Id: I23ece1220559d240f4a48818a4a3d95627e6bf40
Signed-off-by: Doug Hellmann <doug@doughellmann.com>