ensure the validation works when some repositories are not tagged
The validation logic for ensuring that semver is applied when
dependencies are updated did not take into account the fact that
sometimes we do not tag every repository in a deliverable. This change
tracks the previous version associated with each repository separately
so that if repos are added or removed from a deliverable over time the
validation rules can be applied correctly.

Change-Id: I8c911e6fb8a39b1d2ffb12a923a86eb9fe4beb26
Signed-off-by: Doug Hellmann <doug@doughellmann.com>