make sphinxext support eol tags
Fix the sorting for EOL tags so when we collapse the release history
of a project to show it on the web site the sort works.

Also skip over the EOL tags when determining the "latest" release from
the branch. They are still displayed as part of the full tag history.

Change-Id: Iedde30f65bdfa89db0c6697ff031a75350090892
Signed-off-by: Doug Hellmann <doug@doughellmann.com>