support reading zuul settings from within a repo
As part of the python3 goal for stein we are going to move the zuul
settings out of the project-config repository and into each project
tree. We need to update the validation to look at the files in the
repo being released in addition to project-config when we are looking
for the project-templates to verify the release jobs.

Change-Id: Ib64d59b1b6646d779112c9fcad47ae7d2c3c74d4
Signed-off-by: Doug Hellmann <doug@doughellmann.com>