new_release: work from a release history object
The calculation of the last release was working from release
information from targeted series, and used specific code to
retrieve the data from the previous series if that failed.

In preparation for landing code to look further backwards in
history, refactor that code to retrieve and work from a
release history object containing list of releases for each
considered series.

To preserve current behavior, for the moment, we only retrieve
history for the targeted series and the series before that.

This allows to get rid of mocks in get_last_release() tests and
simplify testing from a history of releases.

Change-Id: I48ff9c46c1e9250fbdfc5dd5344f18c7e046735a