fix the calculation used for finding release history
reversed() must be passed a sequence data structure, not a generator.

We cannot slide the result of reversed() so turn it back into a list
so we can slice it.

Include all of the series starting with and including the one given so
that when it is the final series we can still include it in
calculations.

Change-Id: Idbaab36014271c7a60bec6d8a04fc69d51eb13d1
Signed-off-by: Doug Hellmann <doug@doughellmann.com>