pbr 4.2.0
While Oslo is already in feature freeze, we decided to grant an
exception for this release.  The reason is that it adds a couple of
deprecations so we can remove some undesirable features sooner,
whereas if we have to wait until next release it will add a whole
cycle of support before they can be removed.  The only real new
feature in this release is only intended for non-OpenStack users as
a migration path off one of the deprecated features, so there is
little to no risk to the OpenStack release from it.

Change-Id: Iadc390834197dde366b842857f98912a5a1c8a7c