Use --force when checking out commits
The clone_repo script clones the repo if needed and then checks out the
commit, branch, or master. If a previous checkout has left untracked
files, this will fail to protect those files from being overwritten.

In our automation we do not need to care about untracked files and need
to make sure we have the right point in the tree, so force the checkout
to make sure these untracked files are ignored.

Change-Id: Ia1ec2869aa7a899e987ef43a11f0af627efeea65
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>