rocky: release osc-placment 1.3.0
This release adds support for:
  - allocation candidates
  - new format for allocations dictionary
  - nested resource providers
  - limiting the number allocation candidates
  - requiring traits for allocation candidates

This also changes osc-placement from library => client-library as it
is a plugin for openstackclient. And it removes the stable/rocky
branch that had to be manually deleted because of the
mis-categorization of the library as a non-client. The branch was cut
last week as part of the non-client library freeze and was manually
deleted by the infra team in order for us to do this release.

Change-Id: Ifb97c0aa6f3cabcfbe8d9a2791373c873afca9af