Adjust missed branching process
Sometimes creating the branches immediately for projects can comes as a
surprise and cause a little extra work with backports when they have
critical fixes to include with a requirements FFE. To avoid some of
this, change the branch creation policy to allow it at freeze time but
do not enforce it as a requirement. Then create the branches for any
missed deliverables closer to the end of the cycle.

Change-Id: Ib6f4037c450a00080a5e9a2a7665c6ea4d112ba5
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>