Branch the openstack-ansible roles
Our roles are generally consumed from shas instead of tags,
and there is no reason to tag them.

It was decided earlier this cycle to stop tagging those
openstack-ansible roles, but continue to branch
them according to our usual procedures.
We are therefore following the devstack model for our roles.

The roles following that model (i.e. the roles which are not deprecated
or are not 'master' branch only) will be unified in a separate
deliverable in the openstack/governance repo,
as "openstack-ansible-roles" [1].

[1]: https://review.openstack.org/#/c/589446

Change-Id: I7cc717641b5bf09cd8a9c657848c967566f724c8