Remove python-tripleoclient rocky branch
The stable/rocky branch should not have been created for this since it
is a cycle-trailing client and does not have the same client lib
deadline as other clients. The branch is being manually removed by the
infra team. This removes the branch details from the deliverable file so
it doesn't inadvertently get recreated and so the team can later create
the branch at the point where they are ready for it.

Change-Id: I1f0c3439d67772a74c5d98f9410363eec8ee903f
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>