have list-changes show information about other deliverables in the series
We had a few teams "forget" about one or more of their deliverables
during rocky. Let's update the list-changes output to show the other
deliverables owned by a team in a series, along with their latest
version, as a way to prompt the release team to remind the project
team to release those other things.

Change-Id: I79c8c0ee591d557c07c46282da1f7ab14ec6290c
Signed-off-by: Doug Hellmann <doug@doughellmann.com>