Don't reverse series list in new_release
The list is already sorted such that the order is stein, rocky,
queens, etc. If we reverse it and then look at only everything after
and including stein, we get no results unless there has been a stein
release.

The reversing was originally necessary because we were using
os.listdir to get the list of series, but now that we're reading them
from a yaml file that is already reverse sorted we don't need that.
See d3f7e4e1f5c860072a5da070f06498c0ad0aa160.

Change-Id: Id65de92f89dc02ba076a6b2dfb949a091cf4ce38