re-release packages with failures from twine bug
Twine 1.12.0 has a bug that caused it to report a failure even when
the upload was successful. That resulted in quite a few partially
successful releases, with only wheels uploaded to PyPI, no
announcements, and no tarballs uploaded to tarballs.o.o.

This patch re-tags all of those releases to give us good artifacts.

Change-Id: Iaea87ebe23bcad27d4422f3c477bf2fa37cf5bc5
Signed-off-by: Doug Hellmann <doug@doughellmann.com>