Release openstack/manila 5.0.2 off of stable/pike
There have been a number of vendor driver
bug fixes and fixes to the core, including a
database schema fix to ensure share type
quotas work as expected.

There's a critical bugfix to the
in-tree manila_tempest_plugin that enables
deployers and certification users to not
interrupt their deployment to create manila
resources (default share/group types) in order
to invoke the tests.

Change-Id: Ib3832c2942351240e8566c0c8bfdb109e2456afd