Release openstack/manila 4.0.2 off of stable/ocata
There have been a number of vendor driver
bug fixes and fixes to the core, including
a fix to resolve log spamming with database
deprecation messages.

There's also a critical bugfix to the
in-tree manila_tempest_plugin that enables
deployers and certification users to not
interrupt their deployment to create manila
resources (default share/group types) in order
to invoke the tests.

Change-Id: I03cffe646f4ab91c8bcc2ca790792fffba34e972