Remove driverfixes/ocata branching
This branch was synced up with stable/ocata after we adopted extended
maintenance policies and was manually deleted. To prevent further stable
releases (not that they are now expected for ocata) this removes the
branching from the deliverable file so the automation does not think it
is a new branch request.

Change-Id: I49c7917adbd7662eb2b1f33a4d2dae55af4d280f
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>