Add top level LICENSE file
It is nice to have a top level LICENSE file to help users know which
license is used for the repo. I believe everything in the repo is
licenses to Apache2.

Also as we have a LICENSE how hacking 102 enforces that every code file
needs an explict license so add one to __init__.py

Change-Id: I7609e774c1606f8b4597f45567871bdddedec270
Signed-off-by: Paul Belanger <pabelanger@redhat.com>