Release tosca-parser 1.2.0
This is the Stein milestone-1 release for tosca-parser to release merged
commits since the last release.

See [1] for more details on the new cycle-with-intermediary library
release changes.

[1] http://lists.openstack.org/pipermail/openstack-dev/2018-October/135689.html

We plan on processing this release November 1 unless we are told to hold
off, or earlier if we get a +1 from the PTL and/or release liaison
indicating that it is all clear to proceed.

Change-Id: Id0dc1b2c5023e36194fd7ff01032b347d5834549
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>