Release python-watcherclient 2.2.0
This is the Stein milestone-1 release for python-watcherclient to
release merged commits since the last release.

See [1] for more details on the new cycle-with-intermediary library
release changes.

[1] http://lists.openstack.org/pipermail/openstack-dev/2018-October/135689.html

We plan on processing this release November 1 unless we are told to hold
off, or earlier if we get a +1 from the PTL and/or release liaison
indicating that it is all clear to proceed.

Change-Id: I0f96699bf8cf2b2ad9a9580dfc3cd686180bc51b
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>