Release python-novaclient 11.1.0
This is the Stein milestone-1 release for python-novaclient to
release merged commits since the last release.

See [1] for more details on the new cycle-with-intermediary library
release changes.

[1] http://lists.openstack.org/pipermail/openstack-dev/2018-October/135689.html

We plan on processing this release November 1 unless we are told to hold
off, or earlier if we get a +1 from the PTL and/or release liaison
indicating that it is all clear to proceed.

Change-Id: Ibea9ee607b61b03115428a0bb9a6dd128952d0d8
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>