Release python-glanceclient 2.14.0
This is the Stein milestone-1 release for python-glanceclient to
release merged commits since the last release.

See [1] for more details on the new cycle-with-intermediary library
release changes.

[1] http://lists.openstack.org/pipermail/openstack-dev/2018-October/135689.html

We plan on processing this release November 1 unless we are told to hold
off, or earlier if we get a +1 from the PTL and/or release liaison
indicating that it is all clear to proceed.

Depends-on: https://review.openstack.org/614381
Change-Id: Ibcaa2584ed104ad3a3bbdacb3c6f655f5c07511c
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>