Track molteniron and tenks releases
Tenks and Molteniron are two cycle-independent deliverables produced
by the ironic team, but they were not tracked (tenks because it's
too young, molteniron because it only did one manual release).

Both should be released through openstack/releases soon, so let's
create a file for them.

Note that tenks does not have a specific Launchpad project, so we
assume that tenks bugs are filed under Ironic.

Change-Id: I3a583ffcf0ad750e95f91b0590c17f73e9f59ca2