Initial release of os-resource-classes
os-resource-classes is a library within the nova project that provides
standard symbols for resource classes used by the placement service and
clients thereof. It is very similar to the os-traits library, on which
this file is based.

The library currently uses nova's launchpad, as it's expected to have
few bugs because it has no behaviors, just symbols.

Eventually it will be a part of the placement project, when the
separation is complete.

The release is required by placement iself, which would like to
start using this library as seen in
https://review.openstack.org/#/c/623556/

Change-Id: I18cb98de2286a81c0ed275184b4e189f698a2d9c