Release openstack/placement 0.1.0
This will provide a working release of the placement service,
including all the expected dependencies of this cycle. This is expected
to be useful for packagers and other testers. The release will not have
complete documentation, but will be a fully working implementation of
the service. In fact it ought to work installed from pypi without too
much hassle.

Note that placement is currently a deliverable of the nova team and uses
nova's launchpad to track bugs. This is expeced to change relatively
soon but the switchover has not yet happened.

The depends on is to the zuul settings to add the release-to-pypi job.

Depends-On: https://review.openstack.org/628240
Change-Id: I4beed6daf2ed4ce6723d633ebdeca3253fa75727