Make it clear which term a TC election is for
Due to the timing of the Spring summit during the 'Stein' Release we
will hold 2 TC elections.  This change add the release name to the
headings.

Also the TC election is a Cross-project event so move it into that
section.  As there will be more than one election demote the style

Change-Id: If25ca22a336d8675ee8126bb1de2821e40f81551