Add Train elections to schedule
Add both TC and PTL elections to the schedule.  Due to this cycle being
slight different because of the scheduling of the summit, forum and PTG
at essentially the same time we'll have only a single email deadline and
the confirmation of the TC election dates implies the PTL election
dates.

Depends-On: https://review.openstack.org/629693
Change-Id: Ia5ec78a123fbe73e544cc5557f7ae1ebd7a7197d