Fix WordCompleter import
Since May 2018 [1], the WordCompleter and PathCompleter moved
from prompt_toolkit 's contrib module to prompt_toolkit.completion

Without this patch, it is impossible to use the interactive-release
CLI with the latest versions of prompt_toolkit, which is not freezed
in the requirements.

This fixes the issue by using the latest way to import WordCompleter
and ensuring users will have the latest prompt_toolkit versions
compatible with this change.

[1]: https://github.com/prompt-toolkit/python-prompt-toolkit/commit/8d12c16081a50c8baa6b0b4050a195f798b12331#diff-c25fe0509d70181361304341982e6d0f

Change-Id: Ia9180a10f0aa8f6f8ed25d7fd5e4e3ec39f86714