Release 0.2.0 of os-resource-classes
This release adds the 'PCPU' resource class to the library.

It is used to count the number physical processors which have
been marked to be used by a single guest only. This is relevant
to some high performance workloads. A nova blueprint is under
way to use the new class, but it is also considered relevant
generally.

Change-Id: I6f7e534ada21d10dc5b8b6dd7251186479ff94ab