Switch some Telemetry services to cycle-with-rc
There were several deliverables set as cycle-with-intermediary that have
not been doing intermediary releases. That release model is intended for
deliverables that will be releases multiple times throughout the
development cycle.

With the newer cycle-with-rc release model, teams have the option of not
requiring releases during the cycle and just making sure they are done
near the end of the cycle. See [1] for more details.

This proposes to change a few cycle-with-intermediary deliverables to be
cycle-with-rc. Please confirm this is acceptable for the team and feel
free to ask any questions about this change.

[1] http://lists.openstack.org/pipermail/openstack-discuss/2018-December/000465.html

Change-Id: I9c08bb4f25a0db57ac946ce6204a760f3491fb6e
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>