Switch Zun service to cycle-with-rc
There were several deliverables set as cycle-with-intermediary that have
not been doing intermediary releases. That release model is intended for
deliverables that will be releases multiple times throughout the
development cycle.

With the newer cycle-with-rc release model, teams have the option of not
requiring releases during the cycle and just making sure they are done
near the end of the cycle. See [1] for more details.

This proposes to change the cycle-with-intermediary deliverable to be
cycle-with-rc. Please confirm this is acceptable for the team and feel
free to ask any questions about this change.

[1] http://lists.openstack.org/pipermail/openstack-discuss/2018-December/000465.html

Change-Id: I502cab150ebb1543affe89eb0ad84dbadabd92cb
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>