Release neutron stadium stein beta deliverables
The release model was changed in Stein release and
the neutron team only delivered neutron stein b1 only.
This causes a situation that major versions of neutron and
other stadium projects do not match and distributors are forced
to do some tricks so that neutron stein beta and rocky version
of stadium projects.

To address this situation, this commit cuts a new release of
stein b2 for neutron and stein b1 for stadium projects.

In addition, policy-in-code support in stadium projects have
completed, so it is a good timing to release them as well.

Note that this release may help gate failures of networking-bgpvpn
and networking-bagpipe because their gates assume released beta
deliverables. These projects are wokring on migrating legacy jobs
to cope with this situations but beta deliverables from other
neutron stadium projects would help them. This is why I prepare
beta deliverables just two weeks before Stein-3 milestone.

Change-Id: I42839e3168cd2ae2ee13b1fe96776d3982c58838