Add requirements history
We have static URLs redirecting /constraints/upper/series to the
appropriate git interface.  In the next change in this series we move to
a data-driven model so let's supply the data :)

Even though we have tag history that goes back to Folsom only import
Juno and later because that's when we first added constraints support

This means as the create and delete branches for requirements, like
other projects, we have a central source of truth (other than git) from
which to update the htaccess redirects.

I had update deliverables.py to account for the fact most of the
requirements branches don't have a release.  This does make the table
look a little funny (as the earliest release is the series eol tag :()
but I'm not sure how we can do better there.

Change-Id: Ie8e60dc865ab301539c0bb085a52dd25f3f62edf