Release 0.3.0 of os-resource-classes
This release adds the normalize_name utility method which
can be used by clients of the library to generate a
known-good name of a resource class. The change does not
otherwise impact the library so is of minimal risk to
any existing installs.

Also, the team is updated from placement to nova. The
launchpad (for now) remains the same. Storyboard is
in progress, but not happened yet.

Change-Id: Ib878cc7a4e2df1186493b63b3c3d7dcc50fceda0