Final Oslo releases for Stein
This is largely to clean up any pending doc and test changes in the
Oslo projects. Normally those don't require a release, but because
stable branches are created based on the latest release hash, we need
to release these changes so they will be present on the stable branches.
Otherwise we end up having to backport changes unnecessarily.

Change-Id: I4f3b9ea53f61b869bcc329ea18990795f40a4769